A view object based architecture for printing arbitrary objects
Through clever naming of attributes and delegates, all integration tests
are still green. Orders will have the exact same API as before.

* Move invoice rendering methods to Spree::BookkeepingDocument
* Add view objects to connect Spree Objects with Spree::BookkeepingDocument

Your PDF template should be agnostic about the structure of your
printable object. The view objects achieve that.

* Improve NotImplementeError messages
* Add item and adjustment POROs, invoices controller
* Add order#documents tab, translations
* Fix feature spec: Documents tab instead of print buttons
* Rrename all the partials
* Rename "file" to "pdf"
* Add a working index action to the PdfsController
* Fix specs accordingly
* Remove selecting template feature

Before, you were able to select a template from the backend.

As it is considerable development effort to create a Prawn template,
I think it makes sense to remove this option from the user. Now, there's
a PDF model, and you can create a new PDF by creating it with the underlying
Object as `printable` and a string specifying the template to be used.

* First version of index.html.erb
* Make BookkeepingDocument model searchable
* Move invoice number logic to Spree::BookkeepingDocument
* Add a base view that complains when essential things are not defined on a view
* Mmove increasing invoice numbers to base invoice view
* Tests for deprecated methods on Spree::Order
* Use order number for order packaging slip

rebase me 2 index