Rremove unfinished views, templates and tests
In order to get this Gem out of the door, I decided (for the moment) to remove
views and templates for reimbursement invoice and shipment packaging slips. They
are not done yet, and they fill this PR up with lots of code that is not essential.

Hopefully they'll come back.