Add dependent: :destroy to the document associations
When you delete a `Spree::Order`, you'll end up with invalid documents (the
document won't be able to infer its view class or fetch any of its contents).

That happened here, and it'd be much nicer if the document just went away when the
order in question is destroyed, too.

Of course: Do not destroy completed orders if you don't exactly know what you're doing!